test-requirements.txt
faker
inflect
stringcase
neobunch
toolz
pytest
hypothesis
pytz
pyyaml
capybara-py
table2dicts